// apps/client/src/app/probable-waffle/services/achievement/achievement-type.ts
export enum AchievementType {
  // Progression
  FIRST_STEPS = "first_steps",
  CAMPAIGNER = "campaigner",
  WAR_HERO = "war_hero",

  // Milestones
  FIRST_VICTORY = "first_victory",
  THE_ARCHITECT = "the_architect",
  UNIT_COLLECTOR = "unit_collector",
  HUNDRED_WINS = "hundred_wins",

  // Economy
  RESOURCEFUL = "resourceful",
  MASTER_ECONOMIST = "master_economist",
  ECONOMIC_POWERHOUSE = "economic_powerhouse",

  // Military
  ANNIHILATOR = "annihilator",
  UNSTOPPABLE_FORCE = "unstoppable_force",
  SWIFT_VICTORY = "swift_victory",
  COMEBACK_KING = "comeback_king",

  // Challenge
  SCOUTS_HONOR = "scouts_honor",
  NO_FLY_ZONE = "no_fly_zone",
  DEATH_FROM_ABOVE = "death_from_above",
  TURTLE_POWER = "turtle_power",

  // Secret
  MASTER_TACTICIAN = "master_tactician",
  CLICK_HAPPY = "click_happy"
}